working on controllers. made multiple changes need to discuss with Marjan Hasan
