+++ Server/controllers/ownerController.js
@@ -1,0 +1,5 @@
+async function getOwnerHouses (){
+
+}
+
+module.exports = {getOwnerHouses}
