+++ Server/index.js
@@ -2,17 +2,20 @@ const dotenv = require("dotenv");
 const path = require("path");
 const { getDynamicDirname } = require("./config");
 const express = require("express");
-const mongoose = require("mongoose");
 const cors = require("cors");
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
 const passport = require("passport");
-const saltRounds = 10;
 const app = express();
-
+const checkRole = require("./middleware/checkRole");
+// Import routes
+const houseRoutes = require("./routes/houseRoutes");
+const userRoutes = require("./routes/userRoutes");
+const ownerRoutes = require("./routes/ownerRoutes");
+const miscellaneousRoutes = require("./routes/miscellaneousRoutes");
 // config file
 const envPath = path.resolve(getDynamicDirname(), "config", ".env");
 dotenv.config({ path: envPath });
+// DB connection...
+const connectDb = require("./db/connect");
 
 // Port Declaration
 const port = process.env.PORT || 5000;
@@ -23,23 +26,13 @@ app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use(passport.initialize());
 
-const connectDb = require("./db/connect");
-
-// Import the Mongoose models
-const House = require("./models/houseModel");
-const User = require("./models/userModel");
-
 // Passport JWT Strategy
 require("./config/passport");
-
-// Import routes
-const houseRoutes = require("./routes/houseRoutes");
-const userRoutes = require("./routes/userRoutes");
-const miscellaneousRoutes = require("./routes/miscellaneousRoutes");
 
 // Use routes
 app.use(houseRoutes);
 app.use(userRoutes);
+app.use('/owner', checkRole("owner"), ownerRoutes);
 app.use(miscellaneousRoutes);
 
 // Start the server
