+++ Server/middleware/checkRole.js
@@ -1,0 +1,17 @@
+const {check} = require("express-validator");
+
+function checkRole(role) {
+    return (req, res, next) => {
+        const user = req.user;
+
+        if (user && user.roles.includes(role)) {
+            next();
+        } else {
+            res.status(403).json({
+                messaage: "Access Denied"
+            });
+        }
+    }
+}
+
+module.exports = checkRole;
